Crud

+++ SpecialityMetals_Models/Incoming_Models/Incoming.cs
@@ -11,11 +11,11 @@ namespace Speciality_Metals_Back_End.SpecialityMetals_Models.Incoming_Models
 
         public DateTime? Incoming_Date { get; set; }
 
-        public int? Gross_Weight { get; set; }
+        public decimal? Gross_Weight { get; set; }
 
-        public int? Tare_Weight { get; set; }
+        public decimal? Tare_Weight { get; set; }
 
-        public int? Net_Weight { get; set; }
+        public decimal? Net_Weight { get; set; }
 
         public string? GRV_Number { get; set; }
         public int? SupplierID { get; set; }
