+++ SpecialityMetals_Models/Outgoing/Outgoing.cs
@@ -11,11 +11,11 @@ namespace Speciality_Metals_Back_End.SpecialityMetals_Models.Outgoing
 
         public DateTime? Outgoing_Date { get; set; }
 
-        public int? Gross_Weight { get; set; }
+        public decimal? Gross_Weight { get; set; }
 
-        public int? Tare_Weight { get; set; }
+        public decimal? Tare_Weight { get; set; }
 
-        public int? Net_Weight { get; set; }
+        public decimal? Net_Weight { get; set; }
 
         public string? Del_Note { get; set; }
         public int? SupplierID { get; set; }
